+++ hack/update-codegen.sh
@@ -60,7 +60,7 @@ ${GOPATH}/bin/deepcopy-gen \
   -O zz_generated.deepcopy \
   --go-header-file ${REPO_ROOT_DIR}/hack/boilerplate/boilerplate.go.txt \
   -i github.com/knative/serving/pkg/apis/config \
-  -i github.com/knative/serving/pkg/reconciler/clusteringress/config \
+  -i github.com/knative/serving/pkg/reconciler/ingress/config \
   -i github.com/knative/serving/pkg/reconciler/certificate/config \
   -i github.com/knative/serving/pkg/reconciler/configuration/config \
   -i github.com/knative/serving/pkg/reconciler/revision/config \
