+++ pkg/reconciler/ingress/config/testdata/config-istio.yaml
