+++ pkg/reconciler/ingress/resources/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2018 The Knative Authors
+Copyright 2019 The Knative Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -15,5 +15,5 @@ limitations under the License.
 */
 
 // Package resources holds simple functions for synthesizing child resources from
-// a ClusterIngress resource and any relevant ClusterIngress controller configuration.
+// an Ingress resource and any relevant ClusterIngress controller configuration.
 package resources
