+++ pkg/reconciler/ingress/resources/names/doc.go
@@ -1,11 +1,11 @@
 /*
-Copyright 2018 The Knative Authors
+Copyright 2019 The Knative Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+    https://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,10 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-/*
-
-Package clusteringress implements a kubernetes controller which tracks ClusterIngress resource
-and reconcile VirtualService as its child resource.
-
-*/
-package clusteringress
+// Package names holds simple functions for synthesizing resource names.
+package names
